User trending service. Filtering the trending linear data from slop with NaN. Get user list with rank using trending points.

+++ src/main/java/com/questnr/model/dto/UserWithRankDTO.java
@@ -11,7 +11,7 @@ public class UserWithRankDTO extends UserDTO{
 
     private int totalPostVisits;
 
-    private Long userRank;
+    private Double userRank;
 
     public int getTotalFollowers() {
         return totalFollowers;
@@ -53,11 +53,11 @@ public class UserWithRankDTO extends UserDTO{
         this.totalPostVisits = totalPostVisits;
     }
 
-    public Long getUserRank() {
+    public Double getUserRank() {
         return userRank;
     }
 
-    public void setUserRank(Long userRank) {
+    public void setUserRank(Double userRank) {
         this.userRank = userRank;
     }
 }
